re-fixed bug for second time on multifile.pcap

+++ tests/test-multifile.sh
@@ -3,9 +3,18 @@
 
 . ./test-subs.sh
 
+testlist="1 2 3 4 10 100"
+deldir=yes
 echo multifile test with varying number of maxfds
 
-for maxfds in 1 2 3 4 10 100 
+if test x$1 != x ; then
+  echo Just testing $1
+  testlist=$1
+  deldir=no
+fi
+  
+
+for maxfds in $testlist
 do
   /bin/rm -rf out
   cmd="$TCPFLOW -f $maxfds -o out -X out/report.xml -r multifile_25_21.pcap -a"
@@ -69,7 +78,9 @@ do
   checkmd5 "out/192.168.123.101.04660-110.045.186.224.01120" "dcd18bf7b6572443215154539a37d75c" "363"
   checkmd5 "out/192.168.123.101.04661-110.045.186.224.01120" "d202ebd7c286d1ea4734bdbef69431c6" "323"
   checkmd5 "out/202.043.063.139.00443-192.168.123.101.04591" "722c54c6443119b6c411359b9b7a47c2" "53"
-  /bin/rm -rf out
+  if test $deldir == "yes" ; then
+    /bin/rm -rf out
+  fi
 done
